Rename subnet model file
