+++ src/app/crud/crud_ranges.py
@@ -4,8 +4,8 @@ from sqlalchemy.future import select
 from sqlalchemy.orm import load_only, selectinload
 
 from ..models.openlabs_range_model import OpenLabsRangeModel
-from ..models.openlabs_subnet_model import TemplateSubnetModel
 from ..models.openlabs_vpc_model import OpenLabsVPCModel
+from ..models.template_subnet_model import TemplateSubnetModel
 from ..schemas.template_range_schema import (
     TemplateRangeBaseSchema,
     TemplateRangeID,
