+++ src/app/models/template_subnet_model.py
